+++ src/plone/restapi/tests/test_types.py
@@ -43,13 +43,11 @@ class TestJsonSchemaUtils(TestCase):
             'field1': {
                 'title': u'Foo',
                 'description': u'',
-                'required': True,
                 'type': 'boolean'
             },
             'field2': {
                 'title': u'Bar',
                 'description': u'',
-                'required': True,
                 'type': 'string'
             },
         }
@@ -109,7 +107,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'string',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': u'foobar',
         }
         self.assertEqual(jsonschema, expected)
@@ -128,7 +125,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'string',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'widget': 'textarea',
             'default': u'Lorem ipsum dolor sit amet',
             'minLength': 10,
@@ -148,7 +144,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'boolean',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': False,
         }
         self.assertEqual(jsonschema, expected)
@@ -170,7 +165,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'number',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': 0.5,
         }
         self.assertEqual(jsonschema, expected)
@@ -192,7 +186,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'number',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': 0.5,
         }
         self.assertEqual(jsonschema, expected)
@@ -214,7 +207,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'integer',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': 50,
         }
         self.assertEqual(jsonschema, expected)
@@ -234,7 +226,6 @@ class TestJsonSchemaProviders(TestCase):
             'description': u'My great field',
             'enum': ['foo', 'bar'],
             'enumNames': ['Foo', 'Bar'],
-            'required': True,
             'choices': [('foo', 'Foo'), ('bar', 'Bar')],
         }
         self.assertEqual(jsonschema, expected)
@@ -262,11 +253,9 @@ class TestJsonSchemaProviders(TestCase):
             'minItems': 1,
             'uniqueItems': False,
             'additionalItems': True,
-            'required': True,
             'items': {
                 'type': 'string',
                 'title': u'Text',
-                'required': True,
                 'description': u'Text field',
                 'default': u'Default text',
             }
@@ -286,13 +275,11 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'array',
             'title': u'My field',
             'description': u'',
-            'required': True,
             'uniqueItems': True,
             'additionalItems': True,
             'items': {
                 'title': u'',
                 'description': u'',
-                'required': True,
                 'type': 'integer',
             },
             'default': (1, 2),
@@ -311,13 +298,11 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'array',
             'title': u'My field',
             'description': u'',
-            'required': True,
             'uniqueItems': True,
             'additionalItems': True,
             'items': {
                 'title': u'',
                 'description': u'',
-                'required': True,
                 'type': 'string',
             }
         }
@@ -337,13 +322,11 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'array',
             'title': u'My field',
             'description': u'',
-            'required': True,
             'uniqueItems': True,
             'additionalItems': True,
             'items': {
                 'title': u'',
                 'description': u'',
-                'required': True,
                 'type': 'string',
                 'enum': ['foo', 'bar'],
                 'enumNames': ['Foo', 'Bar'],
@@ -365,18 +348,15 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'object',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'properties': {
                 'field1': {
                     'title': u'Foo',
                     'description': u'',
-                    'required': True,
                     'type': 'boolean'
                 },
                 'field2': {
                     'title': u'Bar',
                     'description': u'',
-                    'required': True,
                     'type': 'string'
                 },
             }
@@ -395,7 +375,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'string',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'widget': 'richtext',
         }
         self.assertEqual(jsonschema, expected)
@@ -412,7 +391,6 @@ class TestJsonSchemaProviders(TestCase):
         expected = {
             'type': 'string',
             'title': u'My field',
-            'required': True,
             'description': u'My great field',
             'default': date(2016, 1, 1),
         }
@@ -429,7 +407,6 @@ class TestJsonSchemaProviders(TestCase):
         expected = {
             'type': 'string',
             'title': u'My field',
-            'required': True,
             'description': u'My great field',
         }
         self.assertEqual(jsonschema, expected)