+++ src/plone/restapi/tests/test_types.py
@@ -43,13 +43,11 @@ class TestJsonSchemaUtils(TestCase):
             'field1': {
                 'title': u'Foo',
                 'description': u'',
-                'required': True,
                 'type': 'boolean'
             },
             'field2': {
                 'title': u'Bar',
                 'description': u'',
-                'required': True,
                 'type': 'string'
             },
         }
@@ -113,7 +111,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'string',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': u'foobar',
         }
         self.assertEqual(jsonschema, expected)
@@ -132,7 +129,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'string',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'widget': 'textarea',
             'default': u'Lorem ipsum dolor sit amet',
             'minLength': 10,
@@ -152,7 +148,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'boolean',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': False,
         }
         self.assertEqual(jsonschema, expected)
@@ -174,7 +169,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'number',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': 0.5,
         }
         self.assertEqual(jsonschema, expected)
@@ -196,7 +190,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'number',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': 0.5,
         }
         self.assertEqual(jsonschema, expected)
@@ -218,7 +211,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'integer',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'default': 50,
         }
         self.assertEqual(jsonschema, expected)
@@ -238,7 +230,6 @@ class TestJsonSchemaProviders(TestCase):
             'description': u'My great field',
             'enum': ['foo', 'bar'],
             'enumNames': ['Foo', 'Bar'],
-            'required': True,
             'choices': [('foo', 'Foo'), ('bar', 'Bar')],
         }
         self.assertEqual(jsonschema, expected)
@@ -266,11 +257,9 @@ class TestJsonSchemaProviders(TestCase):
             'minItems': 1,
             'uniqueItems': False,
             'additionalItems': True,
-            'required': True,
             'items': {
                 'type': 'string',
                 'title': u'Text',
-                'required': True,
                 'description': u'Text field',
                 'default': u'Default text',
             }
@@ -290,13 +279,11 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'array',
             'title': u'My field',
             'description': u'',
-            'required': True,
             'uniqueItems': True,
             'additionalItems': True,
             'items': {
                 'title': u'',
                 'description': u'',
-                'required': True,
                 'type': 'integer',
             },
             'default': (1, 2),
@@ -315,13 +302,11 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'array',
             'title': u'My field',
             'description': u'',
-            'required': True,
             'uniqueItems': True,
             'additionalItems': True,
             'items': {
                 'title': u'',
                 'description': u'',
-                'required': True,
                 'type': 'string',
             }
         }
@@ -341,13 +326,11 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'array',
             'title': u'My field',
             'description': u'',
-            'required': True,
             'uniqueItems': True,
             'additionalItems': True,
             'items': {
                 'title': u'',
                 'description': u'',
-                'required': True,
                 'type': 'string',
                 'enum': ['foo', 'bar'],
                 'enumNames': ['Foo', 'Bar'],
@@ -369,18 +352,15 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'object',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'properties': {
                 'field1': {
                     'title': u'Foo',
                     'description': u'',
-                    'required': True,
                     'type': 'boolean'
                 },
                 'field2': {
                     'title': u'Bar',
                     'description': u'',
-                    'required': True,
                     'type': 'string'
                 },
             }
@@ -399,7 +379,6 @@ class TestJsonSchemaProviders(TestCase):
             'type': 'string',
             'title': u'My field',
             'description': u'My great field',
-            'required': True,
             'widget': 'richtext',
         }
         self.assertEqual(jsonschema, expected)
@@ -416,7 +395,6 @@ class TestJsonSchemaProviders(TestCase):
         expected = {
             'type': 'string',
             'title': u'My field',
-            'required': True,
             'description': u'My great field',
             'default': date(2016, 1, 1),
         }
@@ -433,7 +411,6 @@ class TestJsonSchemaProviders(TestCase):
         expected = {
             'type': 'string',
             'title': u'My field',
-            'required': True,
             'description': u'My great field',
         }
         self.assertEqual(jsonschema, expected)