+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - @types improvments: add `required` and `widget`, support RichText field
   [ebrehault]
 
+- Implements navigation and breadcrumbs components (copied from barcelona-mocks)
+  [ebrehault]
+
 
 1.0a1 (2016-07-14)
 ------------------