+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
+- Tests: Limit available languages to a small set to avoid excessive
+  language lists in response dumps used for documentation.
+  [lgraf]
+
 - @types improvments: add `required` and `widget`, support RichText field
   [ebrehault]