+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
+- Tests: Use `freezegun` to freeze hard to control timestamps in
+  response dumps used for documentation.
+  [lgraf]
+
 - Tests: Limit available languages to a small set to avoid excessive
   language lists in response dumps used for documentation.
   [lgraf]