+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-1.0a2 (unreleased)
+1.0a2 (2016-08-20)
 ------------------
 
 New Features: